changed all ints and doubles to floats in the configs

+++ src/main/kotlin/us/timinc/mc/cobblemon/unchained/api/AbstractActionInfluenceBooster.kt
@@ -36,7 +36,7 @@ abstract class AbstractActionInfluenceBooster(
         pokemon: Pokemon,
         species: ResourceLocation,
         form: String,
-        points: Double,
+        points: Float,
         player: ServerPlayer,
     )
 }
